レイアウトの調整と不要な記述の削除

+++ js/main.js
@@ -1,11 +1,8 @@
-const items = JSON.parse(localStorage.getItem('items')) ?? {}
 if (localStorage.getItem('nextId') === null) localStorage.setItem('nextId', '1')
-
-// console.log(items)
 
 const app = {
   data: () => ({
-    items,
+    items: JSON.parse(localStorage.getItem('items')) ?? {},
     newItem: '',
     editTarget: ''
   }),
@@ -19,27 +16,24 @@ const app = {
       localStorage.setItem('items', JSON.stringify(this.items))
       localStorage.setItem('nextId', nextId)
       this.newItem = ''
-      console.log('added item')
     },
     editItem (id) {
       this.editTarget = { id: id, value: this.items[id] }
-      console.log(`edit id:${id}`)
     },
     updateItem () {
       this.items[this.editTarget.id] = this.editTarget.value
       localStorage.setItem('items', JSON.stringify(this.items))
       this.editTarget = ''
-      console.log('edited item')
     },
-    stopEdit () {
+    cancelEdit () {
       this.editTarget = ''
-      console.log('stop edit')
     },
     deleteItem (id) {
-      delete this.items[id]
-      localStorage.setItem('items', JSON.stringify(this.items))
-      console.log(`deleted id:${id}`)
+      if (confirm('Are you sure to delete?')) {
+        delete this.items[id]
+        localStorage.setItem('items', JSON.stringify(this.items))
+      }
     }
-  },
+  }
 }
 Vue.createApp(app).mount('#app')
